0.83 - Korge 1.12.9 - - includes fixes and optimizations for Sprite and SpriteView - Optimizations for Kotlin native

+++ src/commonMain/kotlin/me/emig/engineEmi/graphics/shapes/Shape.kt
@@ -5,7 +5,7 @@ import com.soywiz.korge.view.*
 abstract class Shape(x : Number, y : Number) : Graphics(true) {
 	init {
 		apply {
-			xy(x,y)
+			xy(x.toDouble(),y.toDouble())
 		}
 	}
